test: split product tests
