+++ services/tests/__init__.py
@@ -1,1 +1,7 @@
-from . import test_shopify_helpers, test_shopify_service, test_shopify_sync
+from . import (
+    test_product_deleter,
+    test_product_exporter,
+    test_shopify_helpers,
+    test_shopify_service,
+    test_shopify_sync,
+)
